apps/api/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
pydantic-settings==2.12.0
python-multipart==0.0.20
sqlalchemy==2.0.44
alembic==1.17.2
psycopg[binary]==3.2.12
boto3==1.40.61
python-jose[cryptography]==3.5.0
piper-tts>=1.2.0
pathvalidate>=3.2.0
pypdf>=6.2.0
